+++ src/test_network.rs
@@ -5,8 +5,11 @@ use std::{
 
 use bevy::prelude::*;
 use bevy_renet::renet::{
-    ChannelConfig, ClientAuthentication, RenetClient, RenetConnectionConfig, RenetServer,
-    ServerAuthentication, ServerConfig,
+    transport::{
+        ClientAuthentication, NetcodeClientTransport, NetcodeServerTransport, ServerAuthentication,
+        ServerConfig,
+    },
+    ChannelConfig, ConnectionConfig, RenetClient, RenetServer,
 };
 
 use super::replication_core::NetworkChannels;
@@ -19,22 +22,28 @@ impl Plugin for TestNetworkPlugin {
         app.add_plugins(MinimalPlugins);
 
         let network_channels = app.world.resource_mut::<NetworkChannels>();
-        let client_channels = network_channels.client_channels();
         let server_channels = network_channels.server_channels();
-
-        app.insert_resource(create_server(
-            server_channels.clone(),
-            client_channels.clone(),
-        ));
-        app.insert_resource(create_client(
-            app.world.resource::<RenetServer>().addr().port(),
-            client_channels,
+        let client_channels = network_channels.client_channels();
+        let (server, server_transport) =
+            create_server(server_channels.clone(), client_channels.clone());
+        let (client, client_transport) = create_client(
+            server_transport.addr().port(),
             server_channels,
-        ));
+            client_channels,
+        );
+
+        app.insert_resource(server)
+            .insert_resource(server_transport)
+            .insert_resource(client)
+            .insert_resource(client_transport);
 
         loop {
             app.update();
-            if app.world.resource::<RenetClient>().is_connected() {
+            if app
+                .world
+                .resource::<NetcodeClientTransport>()
+                .is_connected()
+            {
                 break;
             }
         }
@@ -44,39 +53,46 @@ impl Plugin for TestNetworkPlugin {
 const PROTOCOL_ID: u64 = 0;
 
 fn create_server(
-    send_channels_config: Vec<ChannelConfig>,
-    receive_channels_config: Vec<ChannelConfig>,
-) -> RenetServer {
-    const MAX_CLIENTS: usize = 1;
+    server_channels_config: Vec<ChannelConfig>,
+    client_channels_config: Vec<ChannelConfig>,
+) -> (RenetServer, NetcodeServerTransport) {
+    let server = RenetServer::new(ConnectionConfig {
+        server_channels_config,
+        client_channels_config,
+        ..Default::default()
+    });
+
     let current_time = SystemTime::now()
         .duration_since(SystemTime::UNIX_EPOCH)
         .unwrap();
     let server_addr = SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 0);
     let socket = UdpSocket::bind(server_addr).expect("localhost should be bindable");
-    let local_addr = socket
+    let public_addr = socket
         .local_addr()
         .expect("socket should autodetect local address");
-    let server_config = ServerConfig::new(
-        MAX_CLIENTS,
-        PROTOCOL_ID,
-        local_addr,
-        ServerAuthentication::Unsecure,
-    );
-
-    let connection_config = RenetConnectionConfig {
-        send_channels_config,
-        receive_channels_config,
-        ..Default::default()
+    let server_config = ServerConfig {
+        max_clients: 1,
+        protocol_id: PROTOCOL_ID,
+        public_addr,
+        authentication: ServerAuthentication::Unsecure,
     };
 
-    RenetServer::new(current_time, server_config, connection_config, socket).unwrap()
+    let transport = NetcodeServerTransport::new(current_time, server_config, socket).unwrap();
+
+    (server, transport)
 }
 
 fn create_client(
     port: u16,
-    send_channels_config: Vec<ChannelConfig>,
-    receive_channels_config: Vec<ChannelConfig>,
-) -> RenetClient {
+    server_channels_config: Vec<ChannelConfig>,
+    client_channels_config: Vec<ChannelConfig>,
+) -> (RenetClient, NetcodeClientTransport) {
+    let client = RenetClient::new(ConnectionConfig {
+        server_channels_config,
+        client_channels_config,
+        ..Default::default()
+    });
+
     let current_time = SystemTime::now()
         .duration_since(SystemTime::UNIX_EPOCH)
         .unwrap();
@@ -90,12 +106,7 @@ fn create_client(
         server_addr,
         user_data: None,
     };
+    let transport = NetcodeClientTransport::new(current_time, authentication, socket).unwrap();
 
-    let connection_config = RenetConnectionConfig {
-        send_channels_config,
-        receive_channels_config,
-        ..Default::default()
-    };
-
-    RenetClient::new(current_time, socket, connection_config, authentication).unwrap()
+    (client, transport)
 }